tools/requirements.txt
google-genai
psycopg2-binary
google-cloud-alloydb-connector
pyyaml
openai
requests
